+++ checkpoint/orbax/checkpoint/test_utils.py
@@ -413,6 +413,7 @@ def create_sharded_array(arr, mesh, mesh_axes):
 def print_directory(directory: epath.PathLike, level: int = 0):
   """Prints a directory tree for debugging purposes."""
   directory = epath.Path(directory)
+  assert directory.exists()
   assert directory.is_dir()
   level_str = '..' * level
   if level == 0: